feat(unpaged): init

+++ src/sql/blocks/blocks_hash_or_number_next.sql
@@ -104,7 +104,7 @@ boundary_blocks AS (
 SELECT extract(
     epoch
     FROM time
-  ) AS "time",
+  )::INTEGER AS "time",
   height AS "height",
   hash AS "hash",
   slot AS "slot",
