+++ src/sql/blocks/blocks_latest.sql
@@ -1,7 +1,7 @@
 SELECT extract(
     epoch
     FROM b.time
-  ) AS "time",
+  )::INTEGER AS "time",
   b.block_no AS "height",
   encode(b.hash, 'hex') AS "hash",
   b.slot_no AS "slot",
